# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: Experiment Management Service
  description: This is the API description of the Experiment Management Service of the Organicity project
  termsOfService: none
  contact:
    name: Organicity
    url: http://organicity.eu/
  license:
    name: none
    url: http://none
host: localhost:8081
...
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:  
...
  /experiments/{expId}/mainexperimenter :
    get:
      tags:
        - Experiments
      description: |
        Get the ID of the main experimenter
      operationId: getMainExperimenter
      parameters:
        - name: expId
          in: path
          required: true
          type: string
        - name: Authorization
          in: header
          description: Authorization token
          required: true
          type: string
          format: string
          default: Bearer ...          
          
      responses:
        "200": 
          description: OK
          schema:
            properties:
              mainExperimenter:
                type: string
        "404": 
          description: experiment not found
        "403": 
          description: authorization failure
  /experiments/{expId}/quota:
    get:
      tags:
        - Experiments
      description: |
        Get the overall experiment quota
      operationId: getExpQuota
      parameters:
        - name: expId
          in: path
          required: true
          type: string
        - name: Authorization
          in: header
          description: Authorization token
          required: true
          type: string
          format: string
          default: Bearer ...          
      responses:
        "200": 
          description: OK
          schema:
            properties:
              quota:
                type: number
        "404": 
          description: experiment not found
        "403": 
          description: authorization failure        
  /experiments/{expId}/remainingquota:
    get:
      tags:
        - Experiments
      description: |
        Get the remaining quota of the experiment
      operationId: getExpRemQuota
      parameters:
        - name: expId
          in: path
          required: true
          type: string
        - name: Authorization
          in: header
          description: Authorization token
          required: true
          type: string
          format: string
          default: Bearer ...          
      responses:
        "200": 
          description: OK
          schema:
            properties:
              remainingQuota:
                type: number
        "404": 
          description: experiment not found
        "403": 
          description: authorization failure     
  /experiments/{expId}/decreaseremquota:
    post:
      tags:
        - Experiments
      description: |
        Increa
      operationId: postExpDecRemQuota
      parameters:
        - name: expId
          in: path
          required: true
          type: string
        - name: Authorization
          in: header
          description: Authorization token
          required: true
          type: string
          format: string
          default: Bearer ...          
      responses:
        "200": 
          description: OK
          schema:
            properties:
              remainingQuota:
                type: number
        "404": 
          description: experiment not found
        "409": 
          description: the remaining quota could not be descreased (e.g. it is 0)
        "403": 
          description: authorization failure   
  /experiments/{expId}/increaseremquota:
    post:
      tags:
        - Experiments
      description: |
        Increa
      operationId: postExpIncRemQuota
      parameters:
        - name: expId
          in: path
          required: true
          type: string
        - name: Authorization
          in: header
          description: Authorization token
          required: true
          type: string
          format: string
          default: Bearer ...          
      responses:
        "200": 
          description: OK
          schema:
            properties:
              remainingQuota:
                type: number
        "404": 
          description: experiment not found
        "409": 
          description: the remaining quota could not be increased (e.g. reached max. quota )
        "403": 
          description: authorization failure           
  /allexperiments:
    get:
      tags: 
       - Experiments
      description: |
        Get all the existing experiments
      operationId: allExperGet
      parameters: 
        - name: Authorization
          in: header
          description: Authorization token
          required: true
          type: string
          format: string
          default: Bearer ...
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Experiments'
        "403":
          description: user is not allowed
...
definitions:
  Area:
    properties:
      name:
        type: string
      startDate:
        type: string
        format: date
      endDate:
        type: string
        format: date
      coordinates:
        type: array
        items:
          type: number
      min:
        type: number
      max:
        type: number
      weight:
        type: string
  Experiment:
    required:
      - experimenterIds
      - experimentId
      - name
      - registered
      - startDate
      - endDate
      - status
      - description
    properties:
      experimenterIds:
        type: array
        items:
         type: string
      experimentId:
        type: string
      name:
        type: string
      registered:
        type: string
        format: date
      startDate:
        type: string
        format: date
      endDate:
        type: string
        format: date
      status:
        type: string
        enum: ['stopped', 'running']
      description:
        type: string
      area:
        type: array
        items:
          "$ref": "#/definitions/Area"
...
  Experiments:
    required:
      - experiments
    properties:
      experiments:
        type: array
        items: 
          $ref: '#/definitions/Experiment'        
...